Only allow one extension obj of each type to be created
HAL did not expect multiple calls of GetExtensionsGnssX,
but if it happens, it causes a static callback function
in AGnss to get set to null when the first instance of
the obj gets deleted even if a second instance of AGnss
is created. This causes SUPL to fail because the static
callback function in AGnss is null and stops HAL from
being able to request SUPL into the Android Framework.

Bug: 149378039
Change-Id: Ib1ef7b06b7544e98f39b9af5295824ff6b84b398
CRs-fixed: 2614283
Test: test SUPL pass after system server restart